+++ CHANGELOG.md
@@ -3,15 +3,19 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.4-alpha] - 2021-04-09
+
+- Added `dice-dialog` component
+
 ## [0.0.3-alpha] - 2021-04-09
 
-- Adds `dice-tabs` component
+- Added `dice-tabs` component
 
 ## [0.0.2-alpha] - 2021-04-07
 
-- Adds `dice-dropdown` component
+- Added `dice-dropdown` component
 
 ## [0.0.1-alpha] - 2021-04-06
 
 - Initial release
-- Includes a `dice-textarea` component that grows with the text typed inside.
+- Added `dice-textarea` component